% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapverse-package.R
\docType{package}
\name{tmapverse-package}
\alias{tmapverse}
\alias{tmapverse-package}
\title{tmapverse Meta-Package for Thematic Mapping}
\description{
The tmapverse package attaches a set of packages that are
commonly used for spatial plotting with tmap. It includes tmap and its
extensions (tmap.glyphs, tmap.networks, tmap.cartogram), as well as
supporting spatial data packages (sf, stars, terra), and cols4all for
exploring color palettes. The package is intended for use in thematic mapping
workflows and does not include the full set of packages from the R-spatial ecosystem.
}
\author{
\strong{Maintainer}: Martijn Tennekes \email{mtennekes@gmail.com}

Other contributors:
\itemize{
  \item Olivier Roy [contributor]
}

}
