% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapverse_packages.R
\name{tmapverse_packages}
\alias{tmapverse_packages}
\alias{tmapverse_packages_versions}
\title{List all packages in the tmapverse}
\usage{
tmapverse_packages(include_deps = FALSE, recursive = FALSE)

tmapverse_packages_versions(
  include_deps = FALSE,
  recursive = FALSE,
  repos = getOption("repos")
)
}
\arguments{
\item{include_deps}{Include dependencies?}

\item{recursive}{If \code{include_deps}, should nested dependencies also be included?}

\item{repos}{repository}
}
\value{
vector of package names ( \code{tmapverse_packages}) or a \code{data.frame} of version information (\code{tmapverse_packages_versions})
}
\description{
List all packages in the tmapverse. The function \code{tmapverse_packages} will return a character vector of packages. The function \code{tmapverse_packages_versions} will also return the corresponding versions and those of the latest versions available in the specified repository.
}
\examples{
tmapverse_packages()
tmapverse_packages(include_deps = TRUE)
tmapverse_packages(include_deps = TRUE, recursive = TRUE)

tmapverse_packages_versions()
tmapverse_packages_versions(include_deps = TRUE)
}
